posix: bitmap PMM on arm64, unified heap HAL, job control + ISIG, strict build
- arm64 PMM: replace bump allocator with bitmap (free/reuse pages)
- heap: single HAL in mk/vm/heap.c (static on arm64, demand-mapped
  on x86, 8-byte block align on i386), drop per-arch duplicates
- kernel: -fstack-protector-strong + __stack_chk_fail; host tests
  build with -fsanitize=address,undefined
- waitpid: WCONTINUED + pid 0/-1/<-1 matching; killpg/setsid/setpgid
  syscalls; continued/stopped state wakes parent
- vfs_open: parent lock closing the O_EXCL create race
- tty: ISIG handling (VINTR/VQUIT/VSUSP -> SIGINT/SIGQUIT/SIGTSTP)
- build: -Werror -MMD in Makefile, .clang-format, untrack PLAN.md,
  remove *.bak; harden user-copy, mmap W^X, brk, futex, ahci, gpt

+++ mk/vm/vm_object.c
@@ -418,7 +418,7 @@ int vm_object_user_paged_fault(vm_object_t *obj, uint64_t offset,
         prot |= VM_PROT_WRITE;
 
     uint64_t phys_addr;
-    int ret = vm_pager_send_fault(obj->paging_port, (uint64_t)obj, offset, prot, &phys_addr);
+    int ret = vm_pager_send_fault(obj->paging_port, (uint64_t)(uintptr_t)obj, offset, prot, &phys_addr);
     if (ret < 0)
         return ret;
